Allowing only using wrapper (the old way) - for the filters
