+++ src/app/(private)/(home)/data-sources/[id]/components/ColumnRoleFields.tsx
@@ -1,17 +1,4 @@
-import { Check } from "lucide-react";
-import { useState } from "react";
-
 import CustomMultiSelect from "@/components/forms/CustomMultiSelect";
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/shadcn/ui/command";
-import { DropdownMenuContent } from "@/shadcn/ui/dropdown-menu";
-import { cn } from "@/shadcn/utils";
 import type { RouterOutputs } from "@/services/trpc/react";
 
 export function ColumnRoleFields({
@@ -23,16 +10,12 @@ export function ColumnRoleFields({
   nameColumns: string[];
   setNameColumns: (ncs: string[]) => void;
 }) {
-  const [search, setSearch] = useState("");
-
-  const onSelect = (currentValue: string) => {
+  const onChange = (currentValue: string) => {
     if (nameColumns.some((c) => c === currentValue)) {
       setNameColumns(nameColumns.filter((c) => c !== currentValue));
     } else {
       setNameColumns(nameColumns.concat([currentValue]));
     }
-
-    setSearch("");
   };
 
   return (
@@ -40,40 +23,9 @@ export function ColumnRoleFields({
       id="config-name-columns-multi"
       label="Name columns"
       hint="Select one or more fields to use as labels on the map."
+      allOptions={dataSource?.columnDefs.map((cd) => cd.name)}
       selectedOptions={nameColumns}
-    >
-      <DropdownMenuContent align="start">
-        <Command>
-          <CommandInput
-            value={search}
-            onValueChange={setSearch}
-            placeholder="Search..."
-            className="h-9"
-          />
-          <CommandList>
-            <CommandEmpty>No value found.</CommandEmpty>
-            <CommandGroup>
-              {dataSource?.columnDefs.map((cd) => (
-                <CommandItem
-                  key={cd.name}
-                  value={cd.name}
-                  onSelect={(currentValue: string) => onSelect(currentValue)}
-                >
-                  {cd.name}
-                  <Check
-                    className={cn(
-                      "ml-auto",
-                      nameColumns.some((c) => c === cd.name)
-                        ? "opacity-100"
-                        : "opacity-0",
-                    )}
-                  />
-                </CommandItem>
-              ))}
-            </CommandGroup>
-          </CommandList>
-        </Command>
-      </DropdownMenuContent>
-    </CustomMultiSelect>
+      onChange={onChange}
+    ></CustomMultiSelect>
   );
 }
